backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "openai>=1.0.0",
    "pydantic==2.5.0",
    "python-dotenv>=1.0.0",
    "uvicorn==0.24.0",
]
